requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
spacy
tenacity
redis==5.0.1
